Simple Structure Defined
This commit creates the initial structure for the project. It includes
a simple launch activity with three hardcoded buttons for now. Each button
launches the schedule activity and passes an extra of data into the full
screen activity. The full screen activity autohides UI elements. This is
mostly out of the box. A few minor color changes to sync things up.