Added basic idea for a table layout
This is just a prototype of what it might look like. The next thing we need
to do is make the interviews based on 15 minute increments and then make
an api that sites on top. It might be worthwhile to see if someone has
created a calendar api we can use instead.